fix 64 bit warnings

+++ library/modules/converters/animationconverter/include/tiki/animationconverter/animationconverter.hpp
@@ -15,7 +15,7 @@ namespace tiki
 
 	protected:
 
-		virtual size_t			getConverterRevision() const { return 9u; }
+		virtual uint16			getConverterRevision() const { return 9u; }
 
 		virtual crc32			getInputType() const;
 		virtual crc32			getOutputType() const;
