+++ library/modules/converters/materialconverter/include/tiki/materialconverter/materialconverter.hpp
@@ -14,7 +14,7 @@ namespace tiki
 
 	protected:
 
-		virtual uint	getConverterRevision() const { return 1u; }
+		virtual uint16	getConverterRevision() const { return 1u; }
 
 		virtual crc32	getInputType() const;
 		virtual crc32	getOutputType() const;
